form item done

+++ app/Http/Controllers/AdminItemController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Item;
+use App\Models\ItemCategory;
+use App\Models\Location;
 use Illuminate\Http\Request;
+use Session;
 
 class AdminItemController extends Controller
 {
@@ -12,6 +15,45 @@ class AdminItemController extends Controller
         $items = Item::with(['location', 'itemStatus', 'user', 'itemCategory'])->get();
         return view('adminItem', compact('items'));
     }
+
+    public function createAdminItem(){
+        $categories = ItemCategory::all();
+        $locations = Location::all();
+
+        return view('admin.form', ['categories' => $categories, 'locations' => $locations]);
+    }
+
+    public function insertAdminItem(Request $request)
+    {
+        $request->validate([
+            'item' => 'required|string|max:255',
+            'category' => 'required|exists:item_categories,id',
+            'description' => 'required|string',
+            'image' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048',
+            'location' => 'required|exists:locations,id',
+            'detail_location' => 'required|string|max:255',
+            'time_found' => 'required|date',
+        ]);
+    
+        $timeFoundTimestamp = strtotime($request->time_found);
+    
+        $item = new Item();
+        $item->name = $request->item;
+        $item->item_category_id = $request->category;
+        $item->item_status_id = 2;
+        $item->description = $request->description;
+        $item->image =  $this->uploadImage($request);
+        $item->location_id = $request->location;
+        $item->location_found = $request->detail_location;
+        $item->time_found = date('Y-m-d H:i:s', $timeFoundTimestamp);
+        $item->save();
+    
+        Session::flash('title', 'Item Berhasil Diinput!');
+        Session::flash('message', '');
+        Session::flash('icon', 'success');
+    
+        return redirect()->route('admin.showAdminItem')->with('success', 'Item Berhasil Diinput!');
+    }    
 
     public function editAdminItem(Item $item)
     {
@@ -50,19 +92,19 @@ class AdminItemController extends Controller
                         ->with('success', 'Item deleted successfully.');
     }
 
-    private function uploadImage(Request $request, Item $item)
+    private function uploadImage(Request $request, Item $item = null)
     {
         if ($request->hasFile('image')) {
-            if ($item->image && file_exists(public_path($item->image))) {
+            if ($item && $item->image && file_exists(public_path($item->image))) {
                 unlink(public_path($item->image));
             }
 
             $fileName = time() . '_' . $request->file('image')->getClientOriginalName();
-            $path = $request->file('image')->move(public_path('images'), $fileName);
+            $request->file('image')->move(public_path('images'), $fileName);
             return 'images/' . $fileName;
         }
 
-        return $item->image;
+        return $item ? $item->image : null;
     }
 
     public function updateItemStatus(Request $request, $id)
